Extracted SupportCRC32 to separate code file
- was causing issues because of references from C and C++
- update CMake's accordingly
