+++ targets/CMSIS-OS/ChibiOS/ST_NUCLEO_F091RC/CMakeLists.txt
@@ -95,7 +95,9 @@ include_directories(
     "${CMAKE_CURRENT_BINARY_DIR}"
     ${CMAKE_CURRENT_SOURCE_DIR}
     ${CMAKE_CURRENT_SOURCE_DIR}/common
-
+    ${PROJECT_SOURCE_DIR}/src/CLR/Include
+    ${PROJECT_SOURCE_DIR}/src/HAL/Include
+    
     ${WireProtocol_INCLUDE_DIRS}
     ${CHIBIOS_INCLUDE_DIRS}
     ${ChibiOSnfOverlay_INCLUDE_DIRS}
